clean up utils/index.ts
